+++ cli/events/events.go
@@ -4,14 +4,92 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
+	"os"
+	"sort"
+	"time"
 
+	"github.com/ahmetb/go-linq/v3"
+	"github.com/criblio/scope/libscope"
 	"github.com/criblio/scope/util"
+	"github.com/fatih/color"
 	"github.com/rs/zerolog/log"
 )
 
-// Reader reads a newline delimited JSON documents and sends parsed documents
+// Define color settings
+var nc = color.New
+
+type colorOpts map[string]*color.Color
+
+var (
+	darkColorOpts = colorOpts{
+		"id":      nc(color.FgBlue),
+		"key":     nc(color.FgGreen),
+		"val":     nc(color.FgHiWhite),
+		"_time":   nc(color.FgWhite),
+		"data":    nc(color.FgHiWhite),
+		"proc":    nc(color.FgHiMagenta),
+		"source":  nc(color.FgHiBlack),
+		"default": nc(color.FgWhite),
+	}
+
+	lightColorOpts = colorOpts{
+		"id":      nc(color.FgBlue),
+		"key":     nc(color.FgGreen),
+		"val":     nc(color.FgBlack),
+		"_time":   nc(color.FgBlack),
+		"data":    nc(color.FgHiBlack),
+		"proc":    nc(color.FgHiMagenta),
+		"source":  nc(color.FgHiWhite),
+		"default": nc(color.FgBlack),
+	}
+
+	sourcetypeColors = colorOpts{
+		"console": nc(color.FgHiGreen),
+		"http":    nc(color.FgHiBlue),
+		"fs":      nc(color.FgHiCyan),
+		"net":     nc(color.FgHiRed),
+		"default": nc(color.FgYellow),
+	}
+)
+
+// Event.Body.Data fields to be displayed by default for a given source
+var sourceFields = map[string][]string{
+	"http.req": {"http_host",
+		"http_method",
+		"http_request_content_length",
+		"http_scheme",
+		"http_target"},
+	"http.resp": {"http_host",
+		"http_method",
+		"http_scheme",
+		"http_target",
+		"http_response_content_length",
+		"http_status"},
+	"http-metrics": {"duration",
+		"req_per_sec"},
+	"fs.open": {"file"},
+	"fs.close": {"file",
+		"file_read_bytes",
+		"file_read_ops",
+		"file_write_bytes",
+		"file_write_ops"},
+	"net.open": {"net_peer_ip",
+		"net_peer_port",
+		"net_host_ip",
+		"net_host_port",
+		"net_protocol",
+		"net_transport"},
+	"net.close": {"net_peer_ip",
+		"net_peer_port",
+		"net_bytes_recv",
+		"net_bytes_sent",
+		"net_close_reason",
+		"net_protocol"},
+}
+
+// EventReader reads a newline delimited JSON documents and sends parsed documents
 // to the passed out channel. It exits the process on error.
-func Reader(r io.Reader, initOffset int64, match func(string) bool, out chan map[string]interface{}) (int, error) {
+func EventReader(r io.Reader, initOffset int64, match func(string) bool, out chan libscope.EventBody) (int, error) {
 	br, err := util.NewlineReader(r, match, func(idx int, Offset int64, b []byte) error {
 		event, err := ParseEvent(b)
 		if err != nil {
@@ -21,29 +99,27 @@ func Reader(r io.Reader, initOffset int64, match func(string) bool, out chan map
 			log.Error().Err(err).Msg("error parsing event")
 			return nil
 		}
-		event["id"] = util.EncodeOffset(initOffset + Offset)
+		event.Id = util.EncodeOffset(initOffset + Offset)
 		out <- event
 		return nil
 	})
+
+	// Tell the consumer we're done
 	close(out)
 	return br, err
 }
 
-// ParseEvent returns an event from an array of bytes containing one event
-func ParseEvent(b []byte) (map[string]interface{}, error) {
-	event := map[string]interface{}{}
+// ParseEvent unmarshals event text into an Event, returning only the Event.Body
+func ParseEvent(b []byte) (libscope.EventBody, error) {
+	var event libscope.Event
 	err := json.Unmarshal(b, &event)
 	if err != nil {
-		return event, err
+		return event.Body, err
 	}
-	if eventType, typeExists := event["type"]; typeExists && eventType.(string) == "evt" {
-		return event["body"].(map[string]interface{}), nil
+	if event.Type == "evt" {
+		return event.Body, nil
 	}
-	// config event in event.json is not an error, special case
-	if _, ok := event["info"]; ok {
-		return event, fmt.Errorf("config event")
-	}
-	return event, fmt.Errorf("could not find body in event")
+	return event.Body, fmt.Errorf("could not find body in event")
 }
 
 // EventMatch helps retrieve events from events.json
@@ -58,11 +134,12 @@ type EventMatch struct {
 	Offset      int64
 }
 
-func (em EventMatch) Events(file io.ReadSeeker, in chan map[string]interface{}) error {
+// Events matches events based on EventMatch config
+func (em EventMatch) Events(file io.ReadSeeker, in chan libscope.EventBody) error {
 	var err error
 	if !em.AllEvents && em.Offset == 0 {
 		var err error
-		em.Offset, err = util.FindReverseLineMatchOffset(em.LastN, file, em.Filter())
+		em.Offset, err = util.FindReverseLineMatchOffset(em.LastN, file, em.filter())
 		if err != nil {
 			return fmt.Errorf("Error searching for Offset: %v", err)
 		}
@@ -76,14 +153,15 @@ func (em EventMatch) Events(file io.ReadSeeker, in chan map[string]interface{}) 
 	}
 
 	// Read Events
-	_, err = Reader(file, em.Offset, em.Filter(), in)
+	_, err = EventReader(file, em.Offset, em.filter(), in)
 	if err != nil {
 		return fmt.Errorf("error reading events: %v", err)
 	}
 	return nil
 }
 
-func (em EventMatch) Filter() func(string) bool {
+// filter filters events based on EventMatch config
+func (em EventMatch) filter() func(string) bool {
 	all := []util.MatchFunc{}
 	if !em.AllEvents && em.SkipEvents > 0 {
 		all = append(all, util.MatchSkipN(em.SkipEvents))
@@ -111,4 +189,192 @@ func (em EventMatch) Filter() func(string) bool {
 	return func(line string) bool {
 		return util.MatchAll(all...)(line)
 	}
+}
+
+// PrintEvent prints a single event
+func PrintEvent(in chan libscope.EventBody, jsonOut bool) {
+	enc := json.NewEncoder(os.Stdout)
+	e := <-in
+	if jsonOut {
+		enc.Encode(e)
+	} else {
+		fields := []util.ObjField{
+			{Name: "ID", Field: "id"},
+			{Name: "Proc", Field: "proc"},
+			{Name: "Pid", Field: "pid"},
+			{Name: "Time", Field: "_time", Transform: func(obj interface{}) string { return util.FormatTimestamp(obj.(float64)) }},
+			{Name: "Command", Field: "cmd"},
+			{Name: "Source", Field: "source"},
+			{Name: "Sourcetype", Field: "sourcetype"},
+			{Name: "Host", Field: "host"},
+			{Name: "Channel", Field: "_channel"},
+			{Name: "Data", Field: "data"},
+		}
+		util.PrintObj(fields, e)
+	}
+}
+
+// PrintEvents prints multiple events
+// handles --eval
+// handles --json
+func PrintEvents(in chan libscope.EventBody, fields []string, sortField string, jsonOut, sortReverse, allFields, forceColor bool) {
+	enc := json.NewEncoder(os.Stdout)
+
+	//	var vm *goja.Runtime
+	//	var prog *goja.Program
+	//	if eval != "" {
+	//		vm = goja.New()
+	//		prog, err = goja.Compile("expr", eval, false)
+	//		util.CheckErrSprintf(err, "error compiling JavaScript expression: %v", err)
+	//	}
+	events := make([]libscope.EventBody, 0)
+	for e := range in {
+		/*
+			if eval != "" {
+				for k, v := range e {
+					vm.Set(k, v)
+				}
+				v, err := vm.RunProgram(prog)
+				util.CheckErrSprintf(err, "error evaluating JavaScript expression: %v", err)
+				res := v.Export()
+				switch r := res.(type) {
+				case bool:
+					if !r {
+						continue
+					}
+				default:
+					util.ErrAndExit("error: JavaScript return value is not boolean")
+				}
+			}
+		*/
+		if jsonOut {
+			enc.Encode(e)
+			continue
+		}
+
+		events = append(events, e)
+	}
+
+	// Sort events
+	if sortField != "" {
+		events = sortEvents(events, sortField, sortReverse)
+	}
+
+	// Print events
+	for _, e := range events {
+		out := getEventText(e, forceColor, allFields, fields)
+		fmt.Printf("%s\n", out)
+	}
+}
+
+// color returns the matching color or the default
+func (co colorOpts) color(color string) *color.Color {
+	c := co[color]
+	if c == nil {
+		c = co["default"]
+	}
+	return c
+}
+
+// getEventText selects and colors event fields
+// handles --color
+// handles --fields
+// handles --allfields
+func getEventText(e libscope.EventBody, forceColor, allFields bool, fields []string) string {
+	co := darkColorOpts // Hard coded for now
+	if forceColor {
+		color.NoColor = false
+	}
+
+	out := co.color("id").Sprintf("[%s] ", e.Id)
+	out += co.color("time").Sprintf("%s ", util.FormatTimestamp(e.Time))
+	out += co.color("proc").Sprintf("%s ", e.Proc)
+	out += sourcetypeColors.color(e.SourceType).Sprintf("%s ", e.SourceType)
+	out += co.color("source").Sprintf("%s ", e.Source)
+	if len(fields) > 0 {
+		out += getEventDataText(e.Data, co, fields)
+	} else {
+		out += getEventDataText(e.Data, co, sourceFields[e.Source])
+	}
+
+	if allFields {
+		if len(e.Data) > 0 {
+			out += fmt.Sprintf("[[%s]]", getEventDataText(e.Data, co, []string{}))
+		}
+	}
+	return out
+}
+
+// getEventDataText selects and colors data fields
+func getEventDataText(e map[string]interface{}, co colorOpts, onlyFields []string) string {
+	kv := ""
+
+	// Which keys do we want to display?
+	keys := make([]string, 0, len(e))
+	if len(onlyFields) > 0 {
+		keys = onlyFields
+	} else {
+		for k := range e {
+			keys = append(keys, k)
+		}
+		sort.Strings(keys)
+	}
+
+	// Let's get their values
+	for idx, k := range keys {
+		if _, ok := e[k]; ok {
+			if idx > 0 {
+				kv += co.color("key").Sprint(" ")
+			}
+			kv += co.color("key").Sprintf("%s", k)
+			kv += co.color("val").Sprintf(":%s", util.InterfaceToString(e[k]))
+
+			// Delete from e so allfields knows what didn't get written yet
+			delete(e, k)
+		}
+	}
+	return kv
+}
+
+// sortEvents sorts an array of events
+// handles --sort
+// handles --reverse
+func sortEvents(arr []libscope.EventBody, sortField string, sortReverse bool) []libscope.EventBody {
+	sortFunc := func(i, j interface{}) bool {
+		f1 := util.GetJSONField(i, sortField)
+		if f1 == nil {
+			util.ErrAndExit("field %s does not exist in top level object %s", sortField, util.JSONBytes(i))
+		}
+		f2 := util.GetJSONField(j, sortField)
+		if f2 == nil {
+			util.ErrAndExit("field %s does not exist in top level object %s", sortField, util.JSONBytes(j))
+		}
+		var ret bool
+		switch f1.Value().(type) {
+		case int:
+			ret = f1.Value().(int) > f2.Value().(int)
+		case int64:
+			ret = f1.Value().(int64) > f2.Value().(int64)
+		case string:
+			ret = f1.Value().(string) > f2.Value().(string)
+		case float64:
+			ret = f1.Value().(float64) > f2.Value().(float64)
+		case time.Time:
+			ret = f1.Value().(time.Time).After(f2.Value().(time.Time))
+		case time.Duration:
+			ret = f1.Value().(time.Duration) > f2.Value().(time.Duration)
+		default:
+			util.ErrAndExit("unsupported field type %s", sortField)
+		}
+		if sortReverse {
+			return !ret
+		}
+		return ret
+	}
+
+	sorted := make([]libscope.EventBody, 0)
+	from := linq.From(arr)
+	from = from.Sort(sortFunc)
+	from.ToSlice(&sorted)
+	return sorted
 }
